Fix package name Fix part of Qulice's warnings
