+++ pom.xml
@@ -1,4 +1,19 @@
 <?xml version="1.0" encoding="UTF-8"?>
+<!--
+  ~ Copyright (C) 2018-2021 FusionSoft
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License");
+  ~ You may not use this file except in compliance with the License.
+  ~ You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing,
+  ~ software distributed under the License is distributed on an
+  ~ "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
+  ~ either express or implied.
+  ~
+  ~ See the License for the specific language governing permissions
+  ~ and limitations under the License.
+  -->
 
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
     xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
@@ -33,8 +48,7 @@
 
     <scm>
         <connection>scm:git:git://github.com/rocket-3/unescaped-string.git</connection>
-        <developerConnection>scm:git:git://github.com/rocket-3/unescaped-string.git
-        </developerConnection>
+        <developerConnection>scm:git:git://github.com/rocket-3/unescaped-string.git</developerConnection>
         <url>http://github.com/rocket-3/unescaped-string/tree/master</url>
     </scm>
 
@@ -233,7 +247,7 @@
                     <artifactId>maven-jar-plugin</artifactId>
                     <version>3.0.2</version>
                 </plugin>
-                <plugin>git remote add origin https://github.com/rocket-3/unescaped-string.git
+                <plugin>
                     <artifactId>maven-install-plugin</artifactId>
                     <version>2.5.2</version>
                 </plugin>
