--- src/main/java/io/github/rocket_3/UnescapedJavaChars.java
@@ -1,9 +1,0 @@
-package io.github.rocket_3;
-
-public class UnescapedJavaChars extends CharsEnvelope {
-
-    public UnescapedJavaChars(final CharSequence chars) {
-        super(new UnescapeJavaChars().apply(chars));
-    }
-
-}
