--- src/main/java/io/github/rocket_3/UnescapedYamlChars.java
@@ -1,9 +1,0 @@
-package io.github.rocket_3;
-
-public class UnescapedYamlChars extends CharsEnvelope {
-
-    public UnescapedYamlChars(final CharSequence origin) {
-        super(new UnescapeYamlChars().apply(origin));
-    }
-
-}
