+++ src/main/java/ru/fusionsoft/unescaped/package-info.java
@@ -1,0 +1,21 @@
+/*
+ * Copyright (C) 2018-2021 FusionSoft
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * You may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
+ * either express or implied.
+ *
+ * See the License for the specific language governing permissions
+ * and limitations under the License.
+ */
+/**
+ * {@link java.util.function.Function}
+ *  of {@link java.lang.CharSequence}, {@link java.lang.CharSequence} variants of solution
+ *  and {@link java.lang.CharSequence} iface wrappers.
+ */
+package ru.fusionsoft.unescaped;
