--- src/test/java/io/github/rocket_3/UnescapedJavaCharsTest.java
@@ -1,18 +1,0 @@
-package io.github.rocket_3;
-
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-
-class UnescapedJavaCharsTest {
-
-    @Test
-    public void rendersExpectedText() {
-        Assertions.assertEquals(
-            "1 \n \\ \" \063 \u0224",
-            new UnescapedJavaChars(
-                "1 \\n \\ \\\" \\063 \\u0224"
-            ).toString()
-        );
-    }
-
-}
