+++ src/test/java/ru/fusionsoft/unescaped/UnescapedJavaCharsTest.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright (C) 2018-2021 FusionSoft
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * You may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
+ * either express or implied.
+ *
+ * See the License for the specific language governing permissions
+ * and limitations under the License.
+ */
+package ru.fusionsoft.unescaped;
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+
+/**
+ * The tests for {@link UnescapedJavaChars}.
+ * @since 0.1
+ */
+class UnescapedJavaCharsTest {
+
+    @Test
+    public void rendersExpectedText() {
+        Assertions.assertEquals(
+            "1 \n \\ \" \063 \u0224",
+            new UnescapedJavaChars(
+                "1 \\n \\ \\\" \\063 \\u0224"
+            ).toString()
+        );
+    }
+
+}
